+++ Source/Steam Library Manager/Forms/SettingsView.xaml.cs
@@ -90,5 +90,43 @@ namespace Steam_Library_Manager.Forms
                 Logger.Error(ex);
             }
         }
+
+        private async void UplayDbPathSelector_OnClick(object sender, RoutedEventArgs e)
+        {
+            try
+            {
+                using (var dialog = new System.Windows.Forms.OpenFileDialog())
+                {
+                    dialog.Filter = "Uplay Database File|configurations";
+                    if (dialog.ShowDialog() == System.Windows.Forms.DialogResult.OK)
+                    {
+                        Properties.Settings.Default.UplayDbPath = dialog.FileName;
+
+                        if (File.Exists(Properties.Settings.Default.UplayDbPath))
+                        {
+                            await Functions.Uplay.InitializeUplayDb();
+                        }
+                    }
+                }
+            }
+            catch (Exception ex)
+            {
+                Logger.Error(ex);
+            }
+        }
+
+        private async void UplayDbPathClearButton_OnClick(object sender, RoutedEventArgs e)
+        {
+            try
+            {
+                Properties.Settings.Default.UplayDbPath = string.Empty;
+
+                await Functions.Uplay.InitializeUplayDb();
+            }
+            catch (Exception ex)
+            {
+                Logger.Error(ex);
+            }
+        }
     }
 }